add javax.persistence annotations
