+++ src/main/java/trainer/banana/spring/model/Auction.java
@@ -1,103 +1,128 @@
-package trainer.banana.spring;
+package trainer.banana.spring.model;
+
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.Id;
+
+@Entity
 public class Auction {
-    private int id,amount;
-    private String date,time,bid,status ,winner,item;
+
+	@Id @GeneratedValue
+	private int id;
+
+	private int amount;
+	private String date, time, bid, status, winner, item;
+
 	/**
 	 * @return the id
 	 */
 	public int getId() {
 		return id;
 	}
+
 	/**
 	 * @param id the id to set
 	 */
 	public void setId(int id) {
 		this.id = id;
 	}
+
 	/**
 	 * @return the amount
 	 */
 	public int getAmount() {
 		return amount;
 	}
+
 	/**
 	 * @param amount the amount to set
 	 */
 	public void setAmount(int amount) {
 		this.amount = amount;
 	}
+
 	/**
 	 * @return the date
 	 */
 	public String getDate() {
 		return date;
 	}
+
 	/**
 	 * @param date the date to set
 	 */
 	public void setDate(String date) {
 		this.date = date;
 	}
+
 	/**
 	 * @return the time
 	 */
 	public String getTime() {
 		return time;
 	}
+
 	/**
 	 * @param time the time to set
 	 */
 	public void setTime(String time) {
 		this.time = time;
 	}
+
 	/**
 	 * @return the bid
 	 */
 	public String getBid() {
 		return bid;
 	}
+
 	/**
 	 * @param bid the bid to set
 	 */
 	public void setBid(String bid) {
 		this.bid = bid;
 	}
+
 	/**
 	 * @return the status
 	 */
 	public String getStatus() {
 		return status;
 	}
+
 	/**
 	 * @param status the status to set
 	 */
 	public void setStatus(String status) {
 		this.status = status;
 	}
+
 	/**
 	 * @return the winner
 	 */
 	public String getWinner() {
 		return winner;
 	}
+
 	/**
 	 * @param winner the winner to set
 	 */
 	public void setWinner(String winner) {
 		this.winner = winner;
 	}
+
 	/**
 	 * @return the item
 	 */
 	public String getItem() {
 		return item;
 	}
+
 	/**
 	 * @param item the item to set
 	 */
 	public void setItem(String item) {
 		this.item = item;
-	} 
-
+	}
 
 }
