+++ src/main/java/trainer/banana/spring/model/User.java
@@ -1,0 +1,101 @@
+package trainer.banana.spring.model;
+
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.Id;
+
+@Entity
+public class User {
+
+	@Id
+	@GeneratedValue
+	private int id;
+
+	private int age;
+	private String name, country, preferredLanguage, adress;
+
+	/**
+	 * @return the id
+	 */
+	public int getId() {
+		return id;
+	}
+
+	/**
+	 * @param id the id to set
+	 */
+	public void setId(int id) {
+		this.id = id;
+	}
+
+	/**
+	 * @return the age
+	 */
+	public int getAge() {
+		return age;
+	}
+
+	/**
+	 * @param age the age to set
+	 */
+	public void setAge(int age) {
+		this.age = age;
+	}
+
+	/**
+	 * @return the name
+	 */
+	public String getName() {
+		return name;
+	}
+
+	/**
+	 * @param name the name to set
+	 */
+	public void setName(String name) {
+		this.name = name;
+	}
+
+	/**
+	 * @return the preferredLanguage
+	 */
+	public String getPreferredLanguage() {
+		return preferredLanguage;
+	}
+
+	/**
+	 * @param preferredLanguage the preferredLanguage to set
+	 */
+	public void setPreferredLanguage(String preferredLanguage) {
+		this.preferredLanguage = preferredLanguage;
+	}
+
+	/**
+	 * @return the adress
+	 */
+	public String getAdress() {
+		return adress;
+	}
+
+	/**
+	 * @param adress the adress to set
+	 */
+	public void setAdress(String adress) {
+		this.adress = adress;
+	}
+
+	/**
+	 * @return the country
+	 */
+	public String getCountry() {
+		return country;
+	}
+
+	/**
+	 * @param country the country to set
+	 */
+	public void setCountry(String country) {
+		this.country = country;
+	}
+
+}
